[PDI-16893] TextFileOutput has a dangerous "feature" allowing data to be piped to OS commands. - created TextFileOutputLegacy with capability run as command - removed capability run as command from TextFileOutput
