+++ engine/src/main/java/org/pentaho/di/trans/steps/textfileoutput/TextFileOutputMeta.java
@@ -84,10 +84,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
   @Injection( name = "FILENAME" )
   private String fileName;
 
-  /** Whether to treat this as a command to be executed and piped into */
-  @Injection( name = "RUN_AS_COMMAND" )
-  private boolean fileAsCommand;
-
   /** Whether to push the output into the output of a servlet with the executeTrans Carte/DI-Server servlet */
   @Injection( name = "PASS_TO_SERVLET" )
   private boolean servletOutput;
@@ -212,21 +208,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
 
   public TextFileOutputMeta() {
     super(); // allocate BaseStepMeta
-  }
-
-  /**
-   * @return FileAsCommand
-   */
-  public boolean isFileAsCommand() {
-    return fileAsCommand;
-  }
-
-  /**
-   * @param fileAsCommand
-   *          The fileAsCommand to set
-   */
-  public void setFileAsCommand( boolean fileAsCommand ) {
-    this.fileAsCommand = fileAsCommand;
   }
 
   public boolean isServletOutput() {
@@ -690,7 +671,7 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
     return retval;
   }
 
-  private void readData( Node stepnode, IMetaStore metastore ) throws KettleXMLException {
+  protected void readData( Node stepnode, IMetaStore metastore ) throws KettleXMLException {
     try {
       separator = XMLHandler.getTagValue( stepnode, "separator" );
       if ( separator == null ) {
@@ -736,7 +717,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
       }
 
       fileName = loadSource( stepnode, metastore );
-      fileAsCommand = "Y".equalsIgnoreCase( XMLHandler.getTagValue( stepnode, "file", "is_command" ) );
       servletOutput = "Y".equalsIgnoreCase( XMLHandler.getTagValue( stepnode, "file", "servlet_output" ) );
       doNotOpenNewFileInit =
           "Y".equalsIgnoreCase( XMLHandler.getTagValue( stepnode, "file", "do_not_open_new_file_init" ) );
@@ -826,7 +806,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
     fileFormat = "DOS";
     fileCompression = fileCompressionTypeCodes[FILE_COMPRESSION_TYPE_NONE];
     fileName = "file";
-    fileAsCommand = false;
     servletOutput = false;
     doNotOpenNewFileInit = false;
     extension = "txt";
@@ -912,10 +891,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
     // Replace possible environment variables...
     String retval = space.environmentSubstitute( filename );
     String realextension = space.environmentSubstitute( extension );
-
-    if ( meta.isFileAsCommand() ) {
-      return retval;
-    }
 
     Date now = new Date();
 
@@ -1019,26 +994,7 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
     retval.append( "    " + XMLHandler.addTagValue( "fileNameField", fileNameField ) );
     retval.append( "    " + XMLHandler.addTagValue( "create_parent_folder", createparentfolder ) );
     retval.append( "    <file>" ).append( Const.CR );
-    if ( parentStepMeta != null && parentStepMeta.getParentTransMeta() != null ) {
-      parentStepMeta.getParentTransMeta().getNamedClusterEmbedManager().registerUrl( fileName );
-    }
-    saveSource( retval, fileName );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "is_command", fileAsCommand ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "servlet_output", servletOutput ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "do_not_open_new_file_init", doNotOpenNewFileInit ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "extention", extension ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "append", fileAppended ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "split", stepNrInFilename ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "haspartno", partNrInFilename ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "add_date", dateInFilename ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "add_time", timeInFilename ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "SpecifyFormat", specifyingFormat ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "date_time_format", dateTimeFormat ) );
-
-    retval.append( "      " ).append( XMLHandler.addTagValue( "add_to_result_filenames", addToResultFilenames ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "pad", padded ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "fast_dump", fastDump ) );
-    retval.append( "      " ).append( XMLHandler.addTagValue( "splitevery", splitEvery ) );
+    saveFileOptions( retval );
     retval.append( "    </file>" ).append( Const.CR );
 
     retval.append( "    <fields>" ).append( Const.CR );
@@ -1063,6 +1019,28 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
     retval.append( "    </fields>" ).append( Const.CR );
 
     return retval.toString();
+  }
+
+  protected void saveFileOptions( StringBuilder retval ) {
+    if ( parentStepMeta != null && parentStepMeta.getParentTransMeta() != null ) {
+      parentStepMeta.getParentTransMeta().getNamedClusterEmbedManager().registerUrl( fileName );
+    }
+    saveSource( retval, fileName );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "servlet_output", servletOutput ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "do_not_open_new_file_init", doNotOpenNewFileInit ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "extention", extension ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "append", fileAppended ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "split", stepNrInFilename ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "haspartno", partNrInFilename ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "add_date", dateInFilename ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "add_time", timeInFilename ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "SpecifyFormat", specifyingFormat ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "date_time_format", dateTimeFormat ) );
+
+    retval.append( "      " ).append( XMLHandler.addTagValue( "add_to_result_filenames", addToResultFilenames ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "pad", padded ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "fast_dump", fastDump ) );
+    retval.append( "      " ).append( XMLHandler.addTagValue( "splitevery", splitEvery ) );
   }
 
   @Override
@@ -1090,7 +1068,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
       encoding = rep.getStepAttributeString( id_step, "encoding" );
 
       fileName = loadSourceRep( rep, id_step, metaStore );
-      fileAsCommand = rep.getStepAttributeBoolean( id_step, "file_is_command" );
       servletOutput = rep.getStepAttributeBoolean( id_step, "file_servlet_output" );
       doNotOpenNewFileInit = rep.getStepAttributeBoolean( id_step, "do_not_open_new_file_init" );
       extension = rep.getStepAttributeString( id_step, "file_extention" );
@@ -1155,7 +1132,6 @@ public class TextFileOutputMeta extends BaseStepMeta implements StepMetaInterfac
       rep.saveStepAttribute( id_transformation, id_step, "compression", fileCompression );
       rep.saveStepAttribute( id_transformation, id_step, "encoding", encoding );
       saveSourceRep( rep, id_transformation, id_step, fileName );
-      rep.saveStepAttribute( id_transformation, id_step, "file_is_command", fileAsCommand );
       rep.saveStepAttribute( id_transformation, id_step, "file_servlet_output", servletOutput );
       rep.saveStepAttribute( id_transformation, id_step, "do_not_open_new_file_init", doNotOpenNewFileInit );
       rep.saveStepAttribute( id_transformation, id_step, "file_extention", extension );
