+++ engine/src/main/resources/kettle-steps.xml
@@ -20,6 +20,7 @@
  <step id="FilterRows"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.FilterRows</description> <classname>org.pentaho.di.trans.steps.filterrows.FilterRowsMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Flow</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.FilterRows</tooltip> <iconfile>ui/images/FLT.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Filter+rows</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="JavaFilter"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.JavaFilter</description> <classname>org.pentaho.di.trans.steps.javafilter.JavaFilterMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Flow</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.JavaFilter</tooltip> <iconfile>ui/images/JLT.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Java+Filter</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="TextFileOutput"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.TextFileOutput</description> <classname>org.pentaho.di.trans.steps.textfileoutput.TextFileOutputMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Output</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.TextOutputFile</tooltip> <iconfile>ui/images/TFO.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Text+File+Output</documentation_url> <cases_url/> <forum_url/> </step>
+ <step id="TextFileOutputLegacy"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.TextFileOutputLegacy</description> <classname>org.pentaho.di.trans.steps.textfileoutputlegacy.TextFileOutputLegacyMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Output</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.TextOutputFile</tooltip> <iconfile>ui/images/TFO.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Text+File+Output+Legacy</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="SystemInfo"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.GetSystemInfo</description> <classname>org.pentaho.di.trans.steps.systemdata.SystemDataMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.GetSystemInfo</tooltip> <iconfile>ui/images/SYS.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Get+System+Info</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="StreamLookup"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.StreamLookup</description> <classname>org.pentaho.di.trans.steps.streamlookup.StreamLookupMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Lookup</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.Streamlookup</tooltip> <iconfile>ui/images/SLU.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Stream+Lookup</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="Unique"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.UniqueRows</description> <classname>org.pentaho.di.trans.steps.uniquerows.UniqueRowsMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Transform</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.Uniquerows</tooltip> <iconfile>ui/images/UNQ.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Unique+Rows</documentation_url> <cases_url/> <forum_url/> </step>
