+++ engine/src/main/resources/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -352,6 +352,7 @@ BaseStep.TypeLongDesc.ExcuteSQL=Execute SQL script
 StepMeta.Log.SaveNewStep=STEP SAVE insert general...
 StepMeta.Exception.StepCouldNotBeLoaded=Step with id {0} could not be loaded from the repository\!
 BaseStep.TypeLongDesc.TextFileOutput=Text file output
+BaseStep.TypeLongDesc.TextFileOutputLegacy=Text file output (legacy)
 BaseStep.TypeTooltipDesc.Rowflattener=Flattens consecutive rows based on the order in which they appear in the input stream
 BaseStep.TypeLongDesc.Output=Table output
 BaseStep.TypeLongDesc.TableAgileMart=Table Agile Mart
