+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_en_US.properties
@@ -6,7 +6,6 @@ TextFileOutputDialog.SelectOutputFiles.DialogMessage=Output file(s)\:
 TextFileOutputDialog.MinWidth.Button=\ &Minimal width 
 TextFileOutputDialog.Separator.Label=Separator 
 TextFileOutputDialog.ShowFiles.Button=&Show filename(s)...
-TextFileOutputDialog.FileIsCommand.Label=Run this as command instead? 
 TextFileOutputDialog.SpecifyFormat.Label=Specify Date time format
 TextFileOutputDialog.DialogTitle=Text file output
 TextFileOutputMeta.CheckResult.ExpectedInputError=No input received from other steps\!
