+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_es_AR.properties
@@ -7,7 +7,6 @@ TextFileOutputDialog.MinWidth.Button=\ Anchura &M\u00ednima
 TextFileOutputMeta.CheckResult.FilesNotChecked=No se han seleccionado las especificaciones del fichero.
 TextFileOutputMeta.CheckResult.AllFieldsFound=Todos los campos de salida se encuentran en el flujo de entrada.
 TextFileOutputDialog.Compression.Label=Compresi\u00f3n
-TextFileOutputDialog.FileIsCommand.Label=\u00bfEjectuar como comando?
 TextFileOutputDialog.AddPartnr.Label=\u00bfIncluir n\u00famero partici\u00f3n en nombre fichero?
 TextFileOutputDialog.PrecisionColumn.Column=Precisi\u00f3n
 TextFileOutputDialog.LengthColumn.Column=Longitud
