+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_es_ES.properties
@@ -7,7 +7,6 @@
 TextFileOutputDialog.DialogTitle = Salida Fichero de Texto
 TextFileOutputDialog.FileTab.TabTitle = Fichero
 TextFileOutputDialog.Filename.Label = Nombre Fichero 
-TextFileOutputDialog.FileIsCommand.Label = ¿Ejectuar como comando\?
 TextFileOutputDialog.AddStepnr.Label = ¿Incluir stepnr en nombre fichero\? 
 TextFileOutputDialog.AddPartnr.Label = ¿Incluir número partición en nombre fichero\?
 TextFileOutputDialog.AddDate.Label = ¿Incluir fecha en nombre fichero\?
