+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_fr_FR.properties
@@ -20,7 +20,6 @@ TextFileOutputDialog.Zipped.Label=Compress\u00E9
 TextFileOutputDialog.Header.Label=En-t\u00EAte 
 TextFileOutputDialog.NoFilesFound.DialogMessage=Aucun fichier trouv\u00E9\!  Veuilez svp v\u00E9rifier le nom de fichier/r\u00E9pertoire.
 TextFileOutputDialog.FailedToGetFields.DialogMessage=Impossible de r\u00E9cup\u00E9rer les champs depuis les autres pr\u00E9c\u00E9dentes\!
-TextFileOutputDialog.FileIsCommand.Label=Ex\u00E9cuter en tant que commande
 TextFileOutputDialog.Encoding.Label=Encodage 
 TextFileOutputMeta.CheckResult.FilesNotChecked=Les sp\u00E9cifications du fichier n''ont pas \u00E9t\u00E9 v\u00E9rifi\u00E9es.
 TextFileOutput.Log.ParentFolderNotExist=Le r\u00E9pertoire parent [{0}] est introuvale\!
