+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_it_IT.properties
@@ -6,7 +6,6 @@ TextFileOutputDialog.SelectOutputFiles.DialogMessage=File di output\:
 TextFileOutputDialog.MinWidth.Button=Larghezza &minima
 TextFileOutputDialog.Separator.Label=Separatore 
 TextFileOutputDialog.ShowFiles.Button=&Mostra file...
-TextFileOutputDialog.FileIsCommand.Label=Eseguire questo come comando?
 TextFileOutputDialog.SpecifyFormat.Label=Specifica il formato di data e ora
 TextFileOutputDialog.DialogTitle=Output del file di testo
 TextFileOutputDialog.ServletOutput.Label=Passa output al servlet
