+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_ja_JP.properties
@@ -6,7 +6,6 @@ TextFileOutputDialog.SelectOutputFiles.DialogMessage=\u51fa\u529b\u3059\u308b\u3
 TextFileOutputDialog.MinWidth.Button=\u6700\u5c0f\u6841\u6570\u306b\u3059\u308b(&M) 
 TextFileOutputDialog.Separator.Label=\u533a\u5207\u308a\u6587\u5b57 
 TextFileOutputDialog.ShowFiles.Button=\u30d5\u30a1\u30a4\u30eb\u540d\u306e\u53c2\u7167(&S)
-TextFileOutputDialog.FileIsCommand.Label=\u30b3\u30de\u30f3\u30c9\u7d50\u679c\u306b\u3059\u308b\uff08\u30d5\u30a1\u30a4\u30eb\u51fa\u529b\u3057\u306a\u3044\uff09 
 TextFileOutputDialog.SpecifyFormat.Label=\u65e5\u4ed8\u306e\u8868\u793a\u5f62\u5f0f\u3092\u6307\u5b9a\u3059\u308b
 TextFileOutputDialog.DialogTitle=\u30c6\u30ad\u30b9\u30c8\u30d5\u30a1\u30a4\u30eb\u51fa\u529b
 TextFileOutputMeta.CheckResult.ExpectedInputError=No input received from other steps\!
