+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutput/messages/messages_zh_CN.properties
@@ -6,7 +6,6 @@ TextFileOutputDialog.SelectOutputFiles.DialogMessage=\u8f93\u51fa\u6587\u4ef6\:
 TextFileOutputDialog.MinWidth.Button=\ &\u6700\u5c0f\u5bbd\u5ea6 
 TextFileOutputDialog.Separator.Label=\u5206\u9694\u7b26 
 TextFileOutputDialog.ShowFiles.Button=&\u663e\u793a\u6587\u4ef6\u540d...
-TextFileOutputDialog.FileIsCommand.Label=\u7ed3\u679c\u8f93\u9001\u81f3\u547d\u4ee4\u884c\u6216\u811a\u672c
 TextFileOutputDialog.SpecifyFormat.Label=\u6307\u5b9a\u65e5\u671f\u65f6\u95f4\u683c\u5f0f
 TextFileOutputDialog.DialogTitle=\u6587\u672c\u6587\u4ef6\u8f93\u51fa
 TextFileOutputDialog.ServletOutput.Label=\u8f93\u51fa\u4f20\u9012\u5230servlet
