+++ engine/src/main/resources/org/pentaho/di/trans/steps/textfileoutputlegacy/messages/messages_en_US.properties
@@ -1,0 +1,1 @@
+TextFileOutputDialog.DialogTitle=Text file output (legacy)
