+++ engine/src/test/java/org/pentaho/di/trans/steps/textfileoutput/TextFileOutputMetaInjectionTest.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2017 by Hitachi Vantara : http://www.pentaho.com
+ * Copyright (C) 2002-2018 by Hitachi Vantara : http://www.pentaho.com
  *
  *******************************************************************************
  *
@@ -41,11 +41,6 @@ public class TextFileOutputMetaInjectionTest extends BaseMetadataInjectionTest<T
     check( "FILENAME", new StringGetter() {
       public String get() {
         return meta.getFileName();
-      }
-    } );
-    check( "RUN_AS_COMMAND", new BooleanGetter() {
-      public boolean get() {
-        return meta.isFileAsCommand();
       }
     } );
     check( "PASS_TO_SERVLET", new BooleanGetter() {
