+++ ui/src/main/java/org/pentaho/di/ui/trans/steps/textfileoutput/TextFileOutputDialog.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2017 by Hitachi Vantara : http://www.pentaho.com
+ * Copyright (C) 2002-2018 by Hitachi Vantara : http://www.pentaho.com
  *
  *******************************************************************************
  *
@@ -47,6 +47,7 @@ import org.eclipse.swt.layout.FormData;
 import org.eclipse.swt.layout.FormLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Event;
 import org.eclipse.swt.widgets.FileDialog;
@@ -96,12 +97,8 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
   private TextVar wFilename;
   private FormData fdlFilename, fdbFilename, fdFilename;
 
-  private Label wlFileIsCommand;
-  private Button wFileIsCommand;
-  private FormData fdlFileIsCommand, fdFileIsCommand;
-
   private Label wlServletOutput;
-  private Button wServletOutput;
+  protected Button wServletOutput;
   private FormData fdlServletOutput, fdServletOutput;
 
   private Label wlExtension;
@@ -197,7 +194,7 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
   private TableView wFields;
   private FormData fdFields;
 
-  private TextFileOutputMeta input;
+  protected TextFileOutputMeta input;
 
   private Button wMinWidth;
   private Listener lsMinWidth;
@@ -219,8 +216,8 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
   private Button wSpecifyFormat;
   private FormData fdlSpecifyFormat, fdSpecifyFormat;
 
-  private Label wlCreateParentFolder;
-  private Button wCreateParentFolder;
+  protected Label wlCreateParentFolder;
+  protected Button wCreateParentFolder;
   private FormData fdlCreateParentFolder, fdCreateParentFolder;
 
   private ColumnInfo[] colinf;
@@ -255,7 +252,7 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     formLayout.marginHeight = Const.FORM_MARGIN;
 
     shell.setLayout( formLayout );
-    shell.setText( BaseMessages.getString( PKG, "TextFileOutputDialog.DialogTitle" ) );
+    shell.setText( getDialogTitle() );
 
     int middle = props.getMiddlePct();
     int margin = Const.MARGIN;
@@ -324,28 +321,7 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     fdFilename.right = new FormAttachment( wbFilename, -margin );
     wFilename.setLayoutData( fdFilename );
 
-    // Run this as a command instead?
-    wlFileIsCommand = new Label( wFileComp, SWT.RIGHT );
-    wlFileIsCommand.setText( BaseMessages.getString( PKG, "TextFileOutputDialog.FileIsCommand.Label" ) );
-    props.setLook( wlFileIsCommand );
-    fdlFileIsCommand = new FormData();
-    fdlFileIsCommand.left = new FormAttachment( 0, 0 );
-    fdlFileIsCommand.top = new FormAttachment( wFilename, margin );
-    fdlFileIsCommand.right = new FormAttachment( middle, -margin );
-    wlFileIsCommand.setLayoutData( fdlFileIsCommand );
-    wFileIsCommand = new Button( wFileComp, SWT.CHECK );
-    props.setLook( wFileIsCommand );
-    fdFileIsCommand = new FormData();
-    fdFileIsCommand.left = new FormAttachment( middle, 0 );
-    fdFileIsCommand.top = new FormAttachment( wFilename, margin );
-    fdFileIsCommand.right = new FormAttachment( 100, 0 );
-    wFileIsCommand.setLayoutData( fdFileIsCommand );
-    wFileIsCommand.addSelectionListener( new SelectionAdapter() {
-      public void widgetSelected( SelectionEvent e ) {
-        input.setChanged();
-        enableParentFolder();
-      }
-    } );
+    Control topControl = addAdditionalComponentIfNeed( middle, margin, wFileComp, wFilename );
 
     // Output to servlet (browser, ws)
     //
@@ -354,7 +330,7 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     props.setLook( wlServletOutput );
     fdlServletOutput = new FormData();
     fdlServletOutput.left = new FormAttachment( 0, 0 );
-    fdlServletOutput.top = new FormAttachment( wFileIsCommand, margin );
+    fdlServletOutput.top = new FormAttachment( topControl, margin );
     fdlServletOutput.right = new FormAttachment( middle, -margin );
     wlServletOutput.setLayoutData( fdlServletOutput );
     wServletOutput = new Button( wFileComp, SWT.CHECK );
@@ -362,7 +338,7 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     props.setLook( wServletOutput );
     fdServletOutput = new FormData();
     fdServletOutput.left = new FormAttachment( middle, 0 );
-    fdServletOutput.top = new FormAttachment( wFileIsCommand, margin );
+    fdServletOutput.top = new FormAttachment( topControl, margin );
     fdServletOutput.right = new FormAttachment( 100, 0 );
     wServletOutput.setLayoutData( fdServletOutput );
     wServletOutput.addSelectionListener( new SelectionAdapter() {
@@ -1289,12 +1265,18 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     return stepname;
   }
 
+  protected String getDialogTitle() {
+    return BaseMessages.getString( PKG, "TextFileOutputDialog.DialogTitle" );
+  }
+
+  protected Control addAdditionalComponentIfNeed( int middle, int margin, Composite wFileComp, Composite topComp ) {
+    return topComp;
+  }
+
   protected void setFlagsServletOption() {
     boolean enableFilename = !wServletOutput.getSelection();
     wlFilename.setEnabled( enableFilename );
     wFilename.setEnabled( enableFilename );
-    wlFileIsCommand.setEnabled( enableFilename );
-    wFileIsCommand.setEnabled( enableFilename );
     wlDoNotOpenNewFileInit.setEnabled( enableFilename );
     wDoNotOpenNewFileInit.setEnabled( enableFilename );
     wlCreateParentFolder.setEnabled( enableFilename );
@@ -1444,7 +1426,6 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     if ( input.getFileName() != null ) {
       wFilename.setText( input.getFileName() );
     }
-    wFileIsCommand.setSelection( input.isFileAsCommand() );
     wServletOutput.setSelection( input.isServletOutput() );
     setFlagsServletOption();
     wDoNotOpenNewFileInit.setSelection( input.isDoNotOpenNewFileInit() );
@@ -1545,9 +1526,8 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     dispose();
   }
 
-  private void saveInfoInMeta( TextFileOutputMeta tfoi ) {
+  protected void saveInfoInMeta( TextFileOutputMeta tfoi ) {
     tfoi.setFileName( wFilename.getText() );
-    tfoi.setFileAsCommand( wFileIsCommand.getSelection() );
     tfoi.setServletOutput( wServletOutput.getSelection() );
     tfoi.setCreateParentFolder( wCreateParentFolder.getSelection() );
     tfoi.setDoNotOpenNewFileInit( wDoNotOpenNewFileInit.getSelection() );
@@ -1713,9 +1693,8 @@ public class TextFileOutputDialog extends BaseStepDialog implements StepDialogIn
     wFields.optWidth( true );
   }
 
-  private void enableParentFolder() {
-    wlCreateParentFolder.setEnabled( !wFileIsCommand.getSelection() );
-    wCreateParentFolder.setEnabled( !wFileIsCommand.getSelection() );
+  protected void enableParentFolder() {
+    // it is enabled always in this implementation
   }
 
 }
